Refactored classes for database initialization and connection management:
edited - DatabaseInitializer class
deleted - SessionManager class
